% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_interactions.R
\name{analyze_interactions}
\alias{analyze_interactions}
\title{Analyze single interaction term in linear regression}
\usage{
analyze_interactions(data, interaction_matrix, response_var)
}
\arguments{
\item{data}{- prepared data generated by function prepare_data (only contains binary variables and response variable )}

\item{interaction_matrix}{- Interaction Matrix generated by create_interaction_matrix}

\item{response_var}{- response variable y}
}
\value{
- full_results. A list where each element is a regression coefficient summary for each interaction term in the interaction matrix.
}
\description{
Analyze single interaction term in linear regression
}
\examples{
set.seed(1234)
n <- 100
v1 = rbinom(n, 1, 0.4)
v2 <- rbinom(n, 1, 0.5)
v3 <- rbinom(n, 1, 0.6)
v4 = rnorm(n)
v5 = rbinom(n, 1, 0.6)
interaction_effect1 <- v2 * v3
interaction_effect2 <- v1 * v3
y1 <- 2 * v1 + 2 * v2 + 2 * v3 + 4 * v4 + 3 * v5 + 100 * interaction_effect1 + 100 * interaction_effect2 + rnorm(n,mean=0,sd=1)
data <- data.frame(y = y1, v1 = v1, v2 = v2, v3 = v3, v4 = v4, v5 = v5)
prepared_data <- prepare_data(data, "y")
interaction_matrix <- create_interaction_matrix(prepared_data, "y")


#The way we generate model would be y = v1 + v2 + v3 + v5 + v1 * v2 ,
 #y = v1 + v2 + v3 + v5 + v1 * v3,...,etc
full_results <- analyze_interactions(prepared_data, interaction_matrix, "y")
print(full_results)

}
