% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interaction_matrix.R
\name{create_interaction_matrix}
\alias{create_interaction_matrix}
\title{Create Interaction Matrix}
\usage{
create_interaction_matrix(data, response_var)
}
\arguments{
\item{data}{- prepared data generated by function prepare_data (only contains binary variables and response variable )}

\item{response_var}{- response variable y}
}
\value{
- Interaction Matrix
}
\description{
Create Interaction Matrix
}
\examples{

set.seed(1234)
n <- 100
v1 = rbinom(n, 1, 0.4)
v2 <- rbinom(n, 1, 0.5)
v3 <- rbinom(n, 1, 0.6)
v4 = rnorm(n)
v5 = rbinom(n, 1, 0.6)
interaction_effect1 <- v2 * v3
interaction_effect2 <- v1 * v3
y1 <- 2 * v1 + 2 * v2 + 2 * v3 + 4 * v4 + 3 * v5 + 100 * interaction_effect1 + 100 * interaction_effect2 + rnorm(n,mean=0,sd=1)
data <- data.frame(y = y1, v1 = v1, v2 = v2, v3 = v3, v4 = v4, v5 = v5)
prepared_data <- prepare_data(data, "y")
interaction_matrix <- create_interaction_matrix(prepared_data, "y")
print(head(interaction_matrix))


}
