% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_effects.R
\name{plot_interaction_effects}
\alias{plot_interaction_effects}
\title{Interaction Effects Plots}
\usage{
plot_interaction_effects(data, significant_interactions)
}
\arguments{
\item{data}{- prepared data generated by function prepare_data (only contains binary variables and response variable )}

\item{significant_interactions}{- significant interaction terms' p-values and their names generated by extract_significant}
}
\value{
plots of interaction terms
}
\description{
Interaction Effects Plots
}
\examples{
#' set.seed(1234)
n <- 100
v1 = rbinom(n, 1, 0.4)
v2 <- rbinom(n, 1, 0.5)
v3 <- rbinom(n, 1, 0.6)
v4 = rnorm(n)
v5 = rbinom(n, 1, 0.6)
interaction_effect1 <- v2 * v3
interaction_effect2 <- v1 * v3
y1 <- 2 * v1 + 2 * v2 + 2 * v3 + 4 * v4 + 3 * v5 + 100 * interaction_effect1 + 100 * interaction_effect2 + rnorm(n,mean=0,sd=1)
data <- data.frame(y = y1, v1 = v1, v2 = v2, v3 = v3, v4 = v4, v5 = v5)
prepared_data <- prepare_data(data, "y")
interaction_matrix <- create_interaction_matrix(prepared_data, "y")
full_results <- analyze_interactions(prepared_data, interaction_matrix, "y")
significant_interactions <- extract_significant(full_results)
fit_final_regression <- final_regression(prepared_data, "y")
plots <- plot_interaction_effects(prepared_data, significant_interactions)
#do plots in R-markdown if you want to see all plots in 1 times.
}
