% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Data preparation -- keep response variable(continuous) and all other binary variables}
\usage{
prepare_data(data, response_var)
}
\arguments{
\item{data}{- original data}

\item{response_var}{- response variable y}
}
\value{
cleaned data
}
\description{
Data preparation -- keep response variable(continuous) and all other binary variables
}
\examples{

set.seed(1234)
n <- 100
v1 = rbinom(n, 1, 0.4)
v2 <- rbinom(n, 1, 0.5)
v3 <- rbinom(n, 1, 0.6)
v4 = rnorm(n)
v5 = rbinom(n, 1, 0.6)
interaction_effect1 <- v2 * v3
interaction_effect2 <- v1 * v3
y <- 2 * v1 + 2 * v2 + 2 * v3 + 4 * v4 + 3 * v5 + 100 * interaction_effect1 + 100 * interaction_effect2 + rnorm(n,mean=0,sd=1)
data <- data.frame(y = y, v1 = v1, v2 = v2, v3 = v3, v4 = v4, v5 = v5)
prepared_data <- prepare_data(data, "y")
}
